requirements.txt
pillow
yt-dlp
mutagen
pywebview
requests
flask-cors
flask
